+++ components/planview_leankit/actions/start-task/start-task.mjs
@@ -3,7 +3,7 @@ import planviewLeankit from "../../planview_leankit.app.mjs";
 export default {
   key: "planview_leankit-start-task",
   name: "Start Task",
-  version: "0.0.2",
+  version: "0.0.3",
   annotations: {
     destructiveHint: false,
     openWorldHint: true,